WIP: break up pipeline into chunks
