+++ benchmarks/pipeline.py
@@ -1,7 +1,7 @@
-from time import time
-import numpy as np
 from tifffile import imread
 import pyFAI
+from xpdtools.pipelines.image_preprocess import raw_foreground, \
+    raw_foreground_dark, raw_background, raw_background_dark
 from xpdtools.pipelines.raw_pipeline import *
 # from xpdtools.pipelines.extra import *
 from profilehooks import profile
