+++ xpdtools/pipelines/raw_pipeline.py
@@ -1,85 +1,20 @@
 """Main pipeline for processing images to I(Q) and PDF"""
-import operator as op
 
 import numpy as np
 from skbeam.core.utils import q_to_twotheta
 from streamz_ext import Stream
 
-from xpdtools.calib import img_calibration
-from xpdtools.tools import (load_geo, mask_img, map_to_binner,
+from xpdtools.pipelines.calibration import wavelength, geometry_img_shape
+from xpdtools.pipelines.scattering_correction import pol_corrected_img
+from xpdtools.tools import (mask_img, map_to_binner,
                             fq_getter, pdf_getter, sq_getter, generate_map_bin)
 
 mask_setting = {'setting': 'auto'}
-raw_foreground = Stream(stream_name='raw foreground')
-raw_foreground_dark = Stream(stream_name='raw foreground dark')
-raw_background = Stream(stream_name='raw background')
-raw_background_dark = Stream(stream_name='raw background dark')
-
 # Get the image shape for the binner
-dark_corrected_foreground = (
-    raw_foreground.
-    combine_latest(raw_foreground_dark, emit_on=0).
-    starmap(op.sub)
-)
-dark_corrected_background = (
-    raw_background.
-    combine_latest(raw_background_dark, emit_on=0).
-    starmap(op.sub)
-)
-bg_corrected_img = (
-    dark_corrected_foreground.
-    combine_latest(dark_corrected_background, emit_on=0).
-    starmap(op.sub, stream_name='background corrected img')
-)
-
-# Calibration management
-wavelength = Stream(stream_name='wavelength')
-calibrant = Stream(stream_name='calibrant')
-detector = Stream(stream_name='detector')
-is_calibration_img = Stream(stream_name='Is Calibration')
-geo_input = Stream(stream_name='geometry')
-gated_cal = (
-    bg_corrected_img.
-    combine_latest(is_calibration_img, emit_on=0).
-    filter(lambda a: bool(a[1])).
-    pluck(0, stream_name='Gate calibration'))
-
-gen_geo_cal = (
-    gated_cal.
-    combine_latest(wavelength,
-                   calibrant,
-                   detector, emit_on=0).
-    starmap(img_calibration)
-)
-
-gen_geo = gen_geo_cal.pluck(1)
-
-geometry = (
-    geo_input.combine_latest(is_calibration_img, emit_on=0).
-    filter(lambda a: not bool(a[1])).
-    pluck(0, stream_name='Gate calibration').
-    map(load_geo).
-    union(gen_geo, stream_name='Combine gen and load cal'))
-
-# Image corrections
-img_shape = (bg_corrected_img.
-             map(np.shape).
-             unique(history=1))
-geometry_img_shape = geometry.zip_latest(img_shape)
 
 # Only create map and bins (which is expensive) when needed (new calibration)
 map_res = geometry_img_shape.starmap(generate_map_bin)
 cal_binner = (map_res.starmap(map_to_binner))
-
-polarization_array = (
-    geometry_img_shape.
-    starmap(lambda geo, shape, polarization_factor: geo.polarization(
-        shape, polarization_factor), .99))
-
-pol_correction_combine = (
-    bg_corrected_img
-    .combine_latest(polarization_array, emit_on=bg_corrected_img))
-pol_corrected_img = pol_correction_combine.starmap(op.truediv)
 
 # emit on img so we don't propagate old image data
 # note that the pol_corrected_img has touched the geometry and so always comes
